// bot-challenge/src/enums/step/type.ts
export enum StepType {
  ElementSelection = 'elemSelect',
  PredictionSelection = 'predictSelect',
  ActionSelection = 'actionSelect',
  ActionElementSelection = 'actionElementSelect',
  ActionElementSelected = 'actionElementSelected',
  InputActionElementTyping = 'inputActionElementTyping',
  ApplyClickAction = 'applyClickAction',
  ApplyInputAction = 'applyInputAction',
}